records: centralize dojson receiver
