+++ tests/unit/records/test_records_receivers.py
@@ -24,8 +24,10 @@ from __future__ import absolute_import, division, print_function
 
 import mock
 
+from inspirehep.modules.records.api import InspireRecord
 from inspirehep.modules.records.receivers import (
     dates_validator,
+    earliest_date,
     match_valid_experiments,
     normalize_field_categories,
     populate_inspire_document_type,
@@ -62,6 +64,82 @@ def test_dates_validator_warns_when_date_is_invalid(warning):
 
     warning.assert_called_once_with(
         'MALFORMED: %s value in %s: %s', 'opening_date', 123, 'bar')
+
+
+def test_earliest_date_from_preprint_date():
+    with_preprint_date = InspireRecord({'preprint_date': '2014-05-29'})
+    earliest_date(None, with_preprint_date)
+
+    expected = '2014-05-29'
+    result = with_preprint_date['earliest_date']
+
+    assert expected == result
+
+
+def test_earliest_date_from_thesis_date():
+    with_thesis_date = InspireRecord({
+        'thesis': {'date': '2008'}
+    })
+    earliest_date(None, with_thesis_date)
+
+    expected = '2008'
+    result = with_thesis_date['earliest_date']
+
+    assert expected == result
+
+
+def test_earliest_date_from_thesis_defense_date():
+    with_thesis_defense_date = InspireRecord({
+        'thesis': {'defense_date': '2012-06-01'}
+    })
+    earliest_date(None, with_thesis_defense_date)
+
+    expected = '2012-06-01'
+    result = with_thesis_defense_date['earliest_date']
+
+    assert expected == result
+
+
+def test_earliest_date_from_publication_info_year():
+    with_publication_info_year = InspireRecord({
+        'publication_info': [
+            {'year': '2014'}
+        ]
+    })
+    earliest_date(None, with_publication_info_year)
+
+    expected = '2014'
+    result = with_publication_info_year['earliest_date']
+
+    assert expected == result
+
+
+def test_earliest_date_from_creation_modification_date_creation_date():
+    with_creation_modification_date_creation_date = InspireRecord({
+        'creation_modification_date': [
+            {'creation_date': '2015-11-04'}
+        ]
+    })
+    earliest_date(None, with_creation_modification_date_creation_date)
+
+    expected = '2015-11-04'
+    result = with_creation_modification_date_creation_date['earliest_date']
+
+    assert expected == result
+
+
+def test_earliest_date_from_imprints_date():
+    with_imprints_date = InspireRecord({
+        'imprints': [
+            {'date': '2014-09-26'}
+        ]
+    })
+    earliest_date(None, with_imprints_date)
+
+    expected = '2014-09-26'
+    result = with_imprints_date['earliest_date']
+
+    assert expected == result
 
 
 def test_match_valid_experiments_adds_facet_experiment():
